Refactor - move UserRepository into it's own package
