+++ src/main/java/com/atspickem/backend/BackendApplication.java
@@ -1,5 +1,6 @@
 package com.atspickem.backend;
 
+import com.atspickem.backend.repository.UserRepository;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
