+++ src/main/java/com/atspickem/backend/services/MyUserDetailsService.java
@@ -1,7 +1,7 @@
 package com.atspickem.backend.services;
 
 
-import com.atspickem.backend.UserRepository;
+import com.atspickem.backend.repository.UserRepository;
 import com.atspickem.backend.models.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -9,7 +9,6 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.Optional;
 
 @Service
@@ -27,4 +26,5 @@ public class MyUserDetailsService implements UserDetailsService {
 
         return user.map(MyUserDetails::new).get();
     }
+
 }
